test: Extract response-builder module, add tests
* Extract response related functions from index module into a new module
  called response-builder.

* Add some tests for the public entry point (buildResponse function).
